upd to last commit, everything should be working by now... should be...

+++ com/file.py
@@ -2,6 +2,7 @@ import discord
 import asyncio
 import requests
 import os
+import guildsave
 import dsbot_extensions as ext
 from discord.ext import commands
 from discord.ext.commands.core import has_permissions
@@ -10,6 +11,13 @@ from discord.ext.commands.core import has_permissions
 class File(commands.Cog):
     @commands.command(name="file.upload")
     async def file_upload(self, ctx):
+        guildconf = guildsave.returnGuildJson(
+            str(ctx.message.guild.id))
+
+        if guildconf["HasPermissionToFileCommands"] == False:
+            await ctx.send("This guild has no permission to file commands!")
+            return
+
         await ctx.send(f"ATTACHMENTS:{ctx.message.attachments}")
         await ctx.send("Trying to download file in host...")
 
@@ -27,6 +35,13 @@ class File(commands.Cog):
 
     @commands.command(name="file.download")
     async def file_download(self, ctx, fileName):
+        guildconf = guildsave.returnGuildJson(
+            str(ctx.message.guild.id))
+
+        if guildconf["HasPermissionToFileCommands"] == False:
+            await ctx.send("This guild has no permission to file commands!")
+            return
+
         path = "./uploads/" + fileName
         if os.path.exists(path) == False:
             await ctx.send(f"os.path.exists({path}) returned False! No file to download.")
@@ -36,16 +51,30 @@ class File(commands.Cog):
 
     @commands.command(name="file.getdir")
     async def file_getdir(self, ctx):
+        guildconf = guildsave.returnGuildJson(
+            str(ctx.message.guild.id))
+
+        if guildconf["HasPermissionToFileCommands"] == False:
+            await ctx.send("This guild has no permission to file commands!")
+            return
+
         onlyfiles = [f for f in os.listdir(
             "./uploads/") if os.path.isfile(os.path.join("./uploads/", f))]
         t = "\tFiles in ./uploads/\n"
         for x in onlyfiles:
             t += f"{x}\t\t\t{(os.path.getsize('./uploads/' + x)/1e+6)} MBs\n"
-            #t += x + "\n"
+            # t += x + "\n"
         await ctx.send(f"`{t}`")
 
     @commands.command(name="file.remove")
     @has_permissions(manage_roles=True)
     async def file_remove(self, ctx, filename):
+        guildconf = guildsave.returnGuildJson(
+            str(ctx.message.guild.id))
+
+        if guildconf["HasPermissionToFileCommands"] == False:
+            await ctx.send("This guild has no permission to file commands!")
+            return
+
         os.remove("./uploads/" + filename)
         await ctx.send("os.remove() called!")
